+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/reporter/CheckstyleReporterTest.groovy
@@ -42,27 +42,27 @@ class CheckstyleReporterTest extends AbstractKitTest {
        error.replaceAll("\r", "").contains """
 6 Checkstyle rule violations were found in 2 files
 
-[Misc | NewlineAtEndOfFile] sample.Sample:0
+[Misc | NewlineAtEndOfFile] sample.(Sample.java:0)
   File does not end with a newline.
   http://checkstyle.sourceforge.net/config_misc.html#NewlineAtEndOfFile
 
-[Javadoc | JavadocType] sample.Sample:3
+[Javadoc | JavadocType] sample.(Sample.java:3)
   Missing a Javadoc comment.
   http://checkstyle.sourceforge.net/config_javadoc.html#JavadocType
 
-[Misc | UncommentedMain] sample.Sample:11
+[Misc | UncommentedMain] sample.(Sample.java:11)
   Uncommented main method found.
   http://checkstyle.sourceforge.net/config_misc.html#UncommentedMain
 
-[Misc | NewlineAtEndOfFile] sample.Sample2:0
+[Misc | NewlineAtEndOfFile] sample.(Sample2.java:0)
   File does not end with a newline.
   http://checkstyle.sourceforge.net/config_misc.html#NewlineAtEndOfFile
 
-[Javadoc | JavadocType] sample.Sample2:3
+[Javadoc | JavadocType] sample.(Sample2.java:3)
   Missing a Javadoc comment.
   http://checkstyle.sourceforge.net/config_javadoc.html#JavadocType
 
-[Misc | UncommentedMain] sample.Sample2:11
+[Misc | UncommentedMain] sample.(Sample2.java:11)
   Uncommented main method found.
   http://checkstyle.sourceforge.net/config_misc.html#UncommentedMain