+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Fix multi-module projects reporting (remove duplicate reports)
 * Fix generated html reports links (redundant slash on linux)
 * Fix newline in reporters (use platform specific)
+* Change reporting format so intelliJ IDEA can recognize class reference and show link (eclipse will probably too). Reference line ranges are not shown anymore (pmd, findbugs and column in checkstyle) - always exact line.
 
 ### 1.3.0 (2016-03-30)
 * Update checkstyle 6.14.1 -> 6.17